Little refactor.
